PredictionService/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
kafka-python==2.0.2
py-eureka-client
sqlalchemy
psycopg2-binary==2.9.8
pydantic==1.10.12
python-dotenv==1.0.0
pymysql

# pytest==7.4.2
# pytest-asyncio==0.21.0

# pandas==2.1.1
# numpy==1.26.0
# scikit-learn==1.4.2
